+++ binanceus/GenAutoEncoder.py
@@ -505,7 +505,7 @@ class GenAutoEncoder(IStrategy):
     ############################
 
     # returns (rolling) smoothed version of input column
-    def roll_smooth(self, col) -> np.float:
+    def roll_smooth(self, col) -> float:
         # must return scalar, so just calculate prediction and take last value
 
         # smooth = gaussian_filter1d(col, 4)
@@ -536,7 +536,7 @@ class GenAutoEncoder(IStrategy):
 
         return model
 
-    def roll_get_dwt(self, col) -> np.float:
+    def roll_get_dwt(self, col) -> float:
         # must return scalar, so just calculate prediction and take last value
 
         model = self.get_dwt(col)
