+++ kucoin/DWT_LongShort_RL.py
@@ -270,7 +270,7 @@ class DWT_LongShort_RL(IStrategy):
 
         return model
 
-    def model(self, a: np.ndarray) -> np.float:
+    def model(self, a: np.ndarray) -> float:
         #must return scalar, so just calculate prediction and take last value
         # model = self.dwtModel(np.array(a))
 
@@ -288,7 +288,7 @@ class DWT_LongShort_RL(IStrategy):
         length = len(model)
         return model[length-1]
 
-    def scaledModel(self, a: np.ndarray) -> np.float:
+    def scaledModel(self, a: np.ndarray) -> float:
         #must return scalar, so just calculate prediction and take last value
         # model = self.dwtModel(np.array(a))
 
@@ -303,7 +303,7 @@ class DWT_LongShort_RL(IStrategy):
         length = len(model)
         return model[length-1]
 
-    def scaledData(self, a: np.ndarray) -> np.float:
+    def scaledData(self, a: np.ndarray) -> float:
 
         # scale the data
         standardized = a.copy()
@@ -315,7 +315,7 @@ class DWT_LongShort_RL(IStrategy):
         length = len(scaled)
         return scaled.ravel()[length-1]
 
-    def predict(self, a: np.ndarray) -> np.float:
+    def predict(self, a: np.ndarray) -> float:
 
         # predicts the next value using polynomial extrapolation
