+++ kucoin/FBB_FFT.py
@@ -233,7 +233,7 @@ class FBB_FFT(IStrategy):
     ###################################
 
 
-    def model(self, a: np.ndarray) -> np.float:
+    def model(self, a: np.ndarray) -> float:
         #must return scalar, so just calculate prediction and take last value
         npredict = self.buy_fft_lookahead.value
         model = self.fourierModel(np.array(a))
@@ -265,7 +265,7 @@ class FBB_FFT(IStrategy):
 
         return model
 
-    def scaledModel(self, a: np.ndarray) -> np.float:
+    def scaledModel(self, a: np.ndarray) -> float:
 
         # scale the data
         standardized = a.copy()
@@ -280,7 +280,7 @@ class FBB_FFT(IStrategy):
         length = len(model)
         return model[length-1]
 
-    def scaledData(self, a: np.ndarray) -> np.float:
+    def scaledData(self, a: np.ndarray) -> float:
 
         # scale the data
         standardized = a.copy()
@@ -292,7 +292,7 @@ class FBB_FFT(IStrategy):
         length = len(scaled)
         return scaled.ravel()[length-1]
 
-    def predict(self, a: np.ndarray) -> np.float:
+    def predict(self, a: np.ndarray) -> float:
         #must return scalar, so just calculate prediction and take last value
         npredict = self.fft_lookahead
         # y = self.fourierExtrapolation(np.array(a), 0)
